Move unit tests out of unit/ subfolder.
