+++ extensions/amp-script/0.1/test/test-user-activation-tracker.js
@@ -17,8 +17,8 @@
 import {
   ACTIVATION_TIMEOUT,
   UserActivationTracker,
-} from '../../user-activation-tracker';
-import {htmlFor} from '../../../../../src/static-template';
+} from '../user-activation-tracker';
+import {htmlFor} from '../../../../src/static-template';
 
 describes.realWin('UserActivationTracker', {}, env => {
   let win;
